+++ noodles-vcf/CHANGELOG.md
@@ -4,6 +4,12 @@
 
 ### Changed
 
+  * vcf/async/io/reader/header: Parse header line by line.
+
+    The async header parser can now build a `vcf::Header` by parsing a raw
+    header line by line. This makes it so that it is no longer required to read
+    the entire raw header into memory before parsing.
+
   * vcf/io/writer/record/reference_bases: Resolve IUPAC ambiguity codes
     ([#268]).